+++ src/main/java/org/osgl/exception/ToBeImplemented.java
@@ -1,0 +1,69 @@
+/* 
+ * Copyright (C) 2013 The Java Tool project
+ * Gelin Luo <greenlaw110(at)gmail.com>
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+*/
+package org.osgl.exception;
+
+/*-
+ * #%L
+ * Java Tool
+ * %%
+ * Copyright (C) 2014 - 2017 OSGL (Open Source General Library)
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+
+import org.osgl.util.S;
+
+/**
+ * Could be used when programmer think it is not logic to reach somewhere. 
+ * For example, the default branch of a switch case on an enum value
+ */
+public class ToBeImplemented extends UnsupportedException {
+
+    public ToBeImplemented(){
+        super();
+    }
+
+    public ToBeImplemented(String message){
+        super(message);
+    }
+
+    public ToBeImplemented(String message, Object... args){
+        super(S.fmt(message, args));
+    }
+
+    public ToBeImplemented(Throwable cause){
+        super(cause);
+    }
+
+    public ToBeImplemented(Throwable cause, String message, Object... args) {
+        super(S.fmt(message, args), cause);
+    }
+}